refactor: split DocumentsService into 9 focused sub-services (#463)
Split the 1,488-line monolithic DocumentsService into domain-focused
services under services/ directory, following the auth resolver split
pattern. No circular dependencies; resolver injects sub-services directly.

Services: DocumentCrud, File, Scan, Analysis, Location, Linking,
AbuseReport, ActivityFeed, ScanHistory.
